app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3 